+++ packages/terra-table/CHANGELOG.md
@@ -3,7 +3,12 @@
 ## Unreleased
 
 * Added
+  * Added row selection mode to the table component.
   * Added ability for consumers to specify a custom screen reader message for masked cells.
+
+* Fixed
+  * Fixed the column header background color for the Orion Fusion theme.
+  * Fixed issue where column header cell was receiving tab focus instead of the button for table role.
   
 ## 5.1.1-alpha.0 - (October 25, 2023)