+++ packages/terra-table/CHANGELOG.md
@@ -10,6 +10,7 @@
   * Fixed the column header background color for the Orion Fusion theme.
   * Fixed an issue with row selection integrating with table zebra stripes.
   * Fixed issue where column header cell was receiving tab focus instead of the button for table role.
+  * Fixed column header left border style.
   
 ## 5.1.1-alpha.0 - (October 25, 2023)